Clean-up dead sessions, clean-up code further
* when bokeh server restarts, all sessions are lost, we need to clean-up these
old sessions and not try and update them
* move bokeh code into bokeh_utils allowing for further refactoring
* use new simpler UserSession structure
