+++ happiness/happiness/viz/team.py
@@ -1,22 +1,15 @@
-from contextlib import closing
-
-from bokeh.client import pull_session
 from bokeh.models import ColumnDataSource
 from bokeh.palettes import Spectral4
 
 from .utils import make_plot, make_legend
 
 
-def update_team_data(user, bokeh_session_id):
-    with closing(pull_session(session_id=bokeh_session_id)) as session:
-        teams = user.employee.teams.all()
-        for team in teams:
-            dates, happiness = team.get_team_dates_happiness()
-            source = session.document.select_one(
-                {'type': ColumnDataSource,
-                'name': team.name}
-            )
-            source.data = dict(x=dates, y=happiness)
+def update_team_data(user, session):
+    teams = user.employee.teams.all()
+    for team in teams:
+        dates, happiness = team.get_team_dates_happiness()
+        source = session.document.select_one({'name': team.name})
+        source.data = dict(x=dates, y=happiness)
 
 
 def make_team_plot(user):
