+++ happiness/happiness/viz/teams.py
@@ -1,17 +1,13 @@
-from contextlib import closing
-
-from bokeh.client import pull_session
 from bokeh.models import ColumnDataSource
 from bokeh.palettes import Spectral4
 
 from .utils import make_plot
 
 
-def update_teams_data(user, bokeh_session_id):
-    with closing(pull_session(session_id=bokeh_session_id)) as session:
-        dates, happiness = user.team.get_team_dates_happiness()
-        source = session.document.select_one({'type': ColumnDataSource})
-        source.data = dict(x=dates, y=happiness)
+def update_teams_data(user, session):
+    dates, happiness = user.team.get_team_dates_happiness()
+    source = session.document.select_one({'type': ColumnDataSource})
+    source.data = dict(x=dates, y=happiness)
 
 
 def make_teams_plot(user):
